added authentication checkers to mutations

+++ server/schemas/resolvers.js
@@ -5,28 +5,16 @@ const { Group, User, Topic } = require("../models");
 const resolvers = {
   Query: {
     getUsers: async (parent, args, context) => {
-      if (context.user) {
-        return await User.find({}).populate('groups');
-      }
-      throw AuthenticationError;
+      return await User.find({}).populate('groups');
     },
     getUser: async (parent, args, context) => {
-      if (context.user) {
-        return await User.findOne({ _id: context.user._id }).populate('groups');
-      }
-      throw AuthenticationError;
+      return await User.findOne({ _id: context.user._id }).populate('groups');
     },
     getGroups: async (parent, args, context) => {
-      if (context.user) {
-        return await Group.find({});
-      }
-      throw AuthenticationError;
+      return await Group.find({});
     },
     getGroup: async (parent, args, context) => {
-      if (context.user) {
-        return await Group.findOne({ _id: args.group_id});
-      }
-      throw AuthenticationError;
+      return await Group.findOne({ _id: args.group_id});
     },
     getTopics: async (parent, args, context) => {
       if (context.user) {
@@ -59,38 +47,50 @@ const resolvers = {
       return { token, user };
     },
     addGroup: async (parent, { group_name, group_description, topic_id, skill_level, zoom_link, meet_time, created_by }) => {
-      const group = await Group.create({ group_name, group_description, topic_id, skill_level, zoom_link, meet_time, created_by });
+      if (context.user) {
+        const group = await Group.create({ group_name, group_description, topic_id, skill_level, zoom_link, meet_time, created_by });
 
-      await User.findOneAndUpdate(
-        { _id: created_by },
-        { $addToSet: { groups: group._id } }
-      );
+        await User.findOneAndUpdate(
+          { _id: created_by },
+          { $addToSet: { groups: group._id } }
+        );
 
-      return group;
+        return group;
+      }
+      throw AuthenticationError;
     },
     removeGroup: async (parent, { group_id }) => {
-      const group = await Group.findOneAndDelete({ id_:group_id });
+      if (context.user) {
+        const group = await Group.findOneAndDelete({ id_:group_id });
 
-      await User.updateMany(
-        { groups: group_id },
-        { $pull: { groups: group_id } }
-      );
+        await User.updateMany(
+          { groups: group_id },
+          { $pull: { groups: group_id } }
+        );
 
-      return group;
+        return group;
+      }
+      throw AuthenticationError;
     },
     enroll: async (parent, { user_id, group_id }) => {
-      const user = await User.findOneAndUpdate(
-        { _id: group_id },
-        { $addToSet: { groups: group_id } }
-      );
-      return user;
+      if (context.user) {
+        const user = await User.findOneAndUpdate(
+          { _id: group_id },
+          { $addToSet: { groups: group_id } }
+        );
+        return user;
+      }
+      throw AuthenticationError;
     },
     unEnroll: async (parent, { user_id, group_id }) => {
-      const user = await User.findOneAndUpdate(
-        { _id: group_id },
-        { $pull: { groups: group_id } }
-      );
-      return user;
+      if (context.user) {
+        const user = await User.findOneAndUpdate(
+          { _id: group_id },
+          { $pull: { groups: group_id } }
+        );
+        return user;
+      }
+      throw AuthenticationError;
     },
   },
 };
